requirements.txt
# Core dependencies
pytube>=12.1.0
numpy>=1.20.0
moviepy>=1.0.3
openai-whisper>=20230314
translatepy>=2.3
gTTS>=2.3.1
soundfile>=0.12.1
tqdm>=4.64.1
requests>=2.28.1

# Optional dependencies for advanced features
# Uncomment if needed
# For Bark TTS
# git+https://github.com/suno-ai/bark.git

# For Coqui TTS
# TTS>=0.13.0

# For API service
# fastapi>=0.95.0
# uvicorn>=0.21.1
# pydantic>=1.10.7 